lecture-09/ecom/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-debug-toolbar==4.4.2
djangorestframework==3.15.1
sqlparse==0.5.0
